adopted the setup to work on our dataset

+++ signjoey/search.py
@@ -351,8 +351,8 @@ def beam_search(
         else:
             topk_log_probs = topk_scores.clone()
 
-        # reconstruct beam origin and true word ids from flattened order
-        topk_beam_index = topk_ids.div(decoder.output_size)
+        # reconstruct beam origin and true word ids from flattened order[]
+        topk_beam_index = topk_ids.div(decoder.output_size).long()
         topk_ids = topk_ids.fmod(decoder.output_size)
 
         # map beam_index to batch_index in the flat representation
@@ -417,7 +417,6 @@ def beam_search(
         select_indices = batch_index.view(-1)
         encoder_output = encoder_output.index_select(0, select_indices)
         src_mask = src_mask.index_select(0, select_indices)
-
         if hidden is not None and not transformer:
             if isinstance(hidden, tuple):
                 # for LSTMs, states are tuples of tensors
